+++ app/src/main/java/de/passwordvault/view/activities/PasswordAnalysisActivity.java
@@ -2,15 +2,11 @@ package de.passwordvault.view.activities;
 
 import android.os.Bundle;
 import android.view.View;
-
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.viewpager2.widget.ViewPager2;
-
 import com.google.android.material.tabs.TabLayout;
 import com.google.android.material.tabs.TabLayoutMediator;
-
 import java.util.ArrayList;
 import de.passwordvault.R;
 import de.passwordvault.model.Observable;