+++ app/src/main/java/de/passwordvault/viewmodel/fragments/PasswordAnalysisListViewModel.java
@@ -2,10 +2,12 @@ package de.passwordvault.viewmodel.fragments;
 
 import androidx.lifecycle.ViewModel;
 
+import de.passwordvault.view.activities.PasswordAnalysisListActivity;
+
 
 /**
  * Class implements a view model for the
- * {@link de.passwordvault.view.fragments.PasswordAnalysisListFragment}.
+ * {@link PasswordAnalysisListActivity}.
  *
  * @author  Christian-2003
  * @version 3.4.0